01-docker-terraform/pyproject.toml
[project]
name = "homework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
]
